feat: Add StartUseCase feature which emits start data to update header, analysis, history data in main view model

+++ app/src/main/java/com/example/opgg/data/summoner/SummonerData.kt
@@ -169,7 +169,7 @@ class MatchData {
         val items: List<Item>,
         val needRenew: Boolean,
         val gameId: String,
-        val createDate: Int,
+        val createDate: Long,
         val gameLength: Int,
         val gameType: String,
         val summonerId: String,
